+++ backend/src/main/java/org/camunda/optimize/service/es/schema/type/AlertType.java
@@ -71,7 +71,7 @@ public class AlertType extends StrictTypeMappingCreator {
         .field("type", "boolean")
       .endObject()
       .startObject(THRESHOLD)
-        .field("type", "integer")
+        .field("type", "long")
       .endObject()
 
       .startObject(TRIGGERED)
